requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
psycopg2-binary
python-dotenv
passlib[bcrypt]
pyjwt
qrcode[pil]
reportlab
Pillow
python-multipart
pydantic[email]
pydantic-settings
aiosmtplib
gunicorn
